+++ services/src/main/java/org/keycloak/protocol/oidc/grants/ClientCredentialsGrantType.java
@@ -47,6 +47,8 @@ import org.keycloak.sessions.AuthenticationSessionModel;
 import org.keycloak.sessions.RootAuthenticationSessionModel;
 import org.keycloak.util.TokenUtil;
 
+import java.util.stream.Collectors;
+
 import static org.keycloak.OAuth2Constants.AUTHORIZATION_DETAILS_PARAM;
 
 /**
@@ -97,7 +99,7 @@ public class ClientCredentialsGrantType extends OAuth2GrantTypeBase {
             throw new CorsErrorResponseException(cors, OAuthErrorException.INVALID_REQUEST, "User '" + clientUsername + "' disabled", Response.Status.UNAUTHORIZED);
         }
 
-        String scope = getRequestedScopes();
+        String scope = TokenManager.clientScopePolicy(getRequestedScopes(), clientUser, realm.getClientScopesStream().collect(Collectors.toList()));
 
         RootAuthenticationSessionModel rootAuthSession = new AuthenticationSessionManager(session).createAuthenticationSession(realm, false);
         AuthenticationSessionModel authSession = rootAuthSession.createAuthenticationSession(client);