+++ services/src/main/java/org/keycloak/protocol/oidc/tokenexchange/StandardTokenExchangeProvider.java
@@ -25,6 +25,8 @@ import jakarta.ws.rs.core.Response;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
+
 import org.keycloak.OAuth2Constants;
 import org.keycloak.OAuthErrorException;
 import org.keycloak.authentication.actiontoken.TokenUtils;
@@ -181,7 +183,7 @@ public class StandardTokenExchangeProvider extends AbstractTokenExchangeProvider
 
     // For now, include "scope" parameter as is
     @Override
-    protected String getRequestedScope(AccessToken token, List<ClientModel> targetAudienceClients) {
+    protected String getRequestedScope(AccessToken token, List<ClientModel> targetAudienceClients, UserModel targetUser) {
         String scope = formParams.getFirst(OAuth2Constants.SCOPE);
 
         boolean validScopes;
@@ -199,7 +201,7 @@ public class StandardTokenExchangeProvider extends AbstractTokenExchangeProvider
             throw new CorsErrorResponseException(cors, OAuthErrorException.INVALID_SCOPE, errorMessage, Response.Status.BAD_REQUEST);
         }
 
-        return scope;
+        return TokenManager.clientScopePolicy(scope, targetUser, realm.getClientScopesStream().collect(Collectors.toList()));
     }
 
     @Override